// angular-biometric-designer/src/app/shared/models/helpdesk/issue.detail.ts
export class IssueDetail {
    issue_desc: string = '';
    issue_name: string = '';
    user_submit: string = '';
    date_submit: string = '';
    user_receive: string = '';
    group_receive: string = '';
    prioity: string = '';
    status: string = '';
    branch: string = '';
    scr_reference: string = '';
    filename: string = '';
    fileData: string[] = [];
}